Adding :global option to page_selector helper. Also fixes Issue when using :only option.
The following issue with :only option is now fixed:
If no :only option was passed to the helper, it would select pages with page_layout => nil and public => true. For sure this is an empty array. It does not return an empty collection, because on line 103 the pages_for_select helper returns all pages, also if the passed "pages" collection is empty.

Now:
If no :only option is used, it will NOT be used in the where clause.
Additionally you can pass :global => Boolean as an option to the helper. If true, it will select global pages instead.

The default behavior of this helper is not touched.
Its just refactored and got a new feature.