Use `non-stupid-digest-assets` gem to have non digest versions of compiled assets. Closes #419
As sprocket-rails does not compile the non digests versions of the assets anymore, we have to use the `non-stupid-digest-assets` gem to get them back and to have Tinymce working again.

NOTE: Unless a new version of the gem was released we have to use my fork, to also have the gzip'ed versions of non digest assets.