Removes the creation and dropping of alchemy_users.
Dropping an existing alchemy_users table means loosing data.
To prevent this, this commit removed that migration for dropping the alchemy_users table and also removed the creation of it from the merged migration file.

This change means: The alchemy_users table will never been created for new apps, but will not dropped for upgraded apps that still relies on the existing table and its data.