Changes error on Language from base to default attribute.
If one unchecked the default checkbox of the default language, the page could not be saved, which is absolutely correct. But there was missing a notice for the user. By setting the error object to the default attribute, the notice is now shown for the user in the form.

Fixes language spec