Adds a "missing contents"-link observer, that sends the request via real ajax post.
Unfortunately Rails does not send a remote ajax request with a method set to 'post' as real post request.

This function converts the query string into a real post body. This fixes issues with missing content links having a lots of option values. I.e. Lots of select values in a EssenceSelect.