Adds internal anchor link select to link overlay.
Adds a select box into the internal link tab page of the link overlay.

This select gives the user the option to link to an anchor of current page.

(Also hides the url text field again. It was there only for debugging and has no real value for the user)