make essence_link_view display the link title between <a> tags
the current version of the default partial somewhat counter-
intuitively renders the path or url (that which usually comes
after the href) twice: once after the href and once as the link
text. there is a field link_title - this should be rendered in-
stead. shouldn't it?

keep up the good work! :)