Adds available_locales setter to Alchemy::I18n module.
With this feature developers can dictate which translations are provided for the user in the backend. (I.e. if Alchemy had 10 translations, but the developer wants to provide only 2 of them.)

It can be used with a simple initializer in the host app: `Alchemy::I18n.available_locales = [:en, :de]`