Fix saving page sort order on 3.0
Apparently, the new page sorting algorithm was orinally made for Alchemy 2.6,
which used ```_recursiveItems(li)``` - in 3.0, the signature of that function had changed
to ```_recursiveItems(item)```. However, some of the code *inside* the function
did not appreciate that changed and barfed. I did not find this bug on my own, but had help
from  @acetoxy and @thereisnospoonpt (who wrote the original sorting algorithm patch). Thank you!