Paint the background of dialogs grey
When the screen is y-narrow and a dialog does not get the correct height in
the options hash, all input elements below that height don't have a background.
This is a dirty fix for that. The proper fix would be to adjust the div
```alchemy-dialog-modal``` to the height of its child elements - but I don't know
how to do that...