Refactor picture.rb#default_mask
Returns the default centered image mask for a given size.
If the mask is bigger than the image, the mask is scaled down
so the largest possible part of the image is visible.

The function will always return a Hash where all values are Integer.

At this point, we can't know what Jcrop is going to do to all this,
so Jcrop will have to scale down the default_mask to whatever
it needs.