Move #default_mask to essence_picture.rb
Since ```EssencePicture``` holds the data for cropping and only
```EssencePicture``` can be cropped with a mask, I do think that
the ```#default_mask``` method should be on EssencePicture.

In the test for the ```#crop``` method, ```@picture.default_mask``` was stubbed
with values that cannot possibly arise using the current ```default_mask```
method. This commit sets them to the dimensions of the full image.