Scale the mask no-matter-what and remove minSize for JCrop
Now the mask is also scaled up to include a maximally big area of the image.

Jcrop will ignore the default_mask if it is smaller than the min_Size.
So it makes sense to no set the min_Size option when the image is
smaller than the desired cropped image.