Bump act-as-taggable-on to 3.2.6 and include the necessary migrations
Acts-as-taggable-on 3.0 has just been updated to a version that breaks alchemy
builds with PostgreSQL and MySQL. This makes every pull request fail. This is
an attempt to update it to the latest version and see whether with this PR
the Travis builds go through again.

I've tested tagging locally, and it does work as expected with this and Postgres.

Even if this passes Travis: This has to be tested manually; I'm not at all sure
I'm doing the Right Thing here.