Fixes updating of page urlname path.
The logic of the Page#set_urlname method does not take the root page into account. So if one updates a language root page, then the urlname would include the root page urlname ('root').

This commit also includes an addition to the page info view. It now includes the complete urlname path.