Do not raise, if page layout could not be found.
Everytime a page is missing it's page layout definition we raise an error. That is annyoing for site beeing refactored and keeps endusers with an 500 error screen.

This commit changes the behavior into logging a warning and displaying a meaningful warning to the end user. Also the user is now able to change the page layout or delete the page and not nagged with an error screen anymore.