Do not patch the ApplicationController anymore.
There are very limited usecases that justify the patching of the ApplicationController and therefore load the language and site on every request of every controller.

If you have controllers that loads Alchemy content or uses Alchemy helpers in
the views (i.e. `render_navigation` or `render_elements`) you should

    include Alchemy::ControllerActions

in these controllers.