Prevent PictureController from sending Set-Cookie
The PictureController will send Set-Cookie headers even when it sends
images. This prevents most caches like Varnish from caching the image
even though correct Cache-Control headers are set. Not sending
Set-Cookie should not be a problem for those files as they are never
used to create or keep a session.