Symlink tmp/cache on deployments
When deploying an Alchemy app with capistrano the `tmp/cache` was not shared between releases. This commit symlinks the whole cache folder to the shared directory to make sure existing cache is not gone after the deployment of a new release.

Background: Alchemy's publishing feature relies on the cache. When you publish a page the dependent cache is getting swiped. Not sharing the cache between releases means public pages which have unpublished content were published just after a deployment.