Introduces `on_page_layout` mixin.
This mixin lets you define callbacks that run on specific page layouts (or all) in the instance of the Alchemy::PagesController.

This is useful for additional things you need to do on your Alchemy pages, i.e. set instance variables or redirects if the current page doesn't meet your requirements.