Adds a wrapper partial for breadcrumb renderer.
The breadcrumb is now wrapped inside a div with class breadcrumb. As with any partial this can be overriden by the the host app.

Also renames the spacer partial into separator.

Note: Due to a bug in Rails, we need to remove the spacer_template key from the render call, if option[:separator] is blank.