Normalize Essence Type: Allow Essences outside Alchemy
I have an Event model outside the Alchemy namespace that perfectly
acts_as_essence. However, I can't use it on pages, because the
"normalize_essence_type" method adds the Alchemy namespace even though
its not necessary.

Initially I though about actually removing the method, but that would be
very cumbersome as then all type entries in the elements.yml file would
have to be namespaced, too - so I included a check for whether a class
with that name exists and an instane of that class would act as an
essence.

In a sense, this is a follow-up to c1fa728.