Adds nestable elements feature.
A nestable element is an element nested inside another element. This
is much more flexible then cells and will replace the cells (as fixed elements)
in future releases.

To create a nestable element define it as you would define a normal element
and add it's name into the `nestable_elements` collection of the parent's element definition
of your `page_layouts.yml` file.