Revert "Execute bundle install right before alchemy:install is called"
The `rails new` command already runs `bundle install` for us (You need to explicitely disable it with `--skip-bundle`, what we don't do).

So the error mentioned in 8e0bc68 must have another reason. Maybe it's `spring`.

In 76b90cf we ensure that spring is disabled while installing, so hopefully the install problems are gone now.