Fixes resource search form.
The search from reset button (x) now only appears if the search field query is present.

Added support for passing in additional query fields, so the will be passed together with the search field query. Nice if you have an additional select inside the toolbar, which query should be passed in.