Adds locale select to languages admin form
In order to fix issues like #831 we add a locale field to languages and validate the presence of it.

So that, if we can't find a locale file (from `I18n.available_locales`) for given language code, users are able to assign locales to alchemy languages to
avoid "missing locale" errors in Rails apps with `config.enforce_available_locales = true`.