WIP: Adds locale select to languages admin form.
In order to fix issues like #831 we add a locale field to languages.
So that users are able to assign locale files to alchemy languages to
avoid errors with `config.enforce_available_locales = true`.

TODO: Use all `I18n.available_locales` instead of only the files present
in the local app.