Always pass on search and page parameters through resource controller actions
My client has big resources through which they need to search, then
modify an entry, then continue with the search. This commit keeps the search
parameters constant during that kind of workflow.

Add helper method `current_location_params` and use it for redirects

In order to always return to the index you left when deleting or updating records,
this commit adds a `current_location_params` helper and uses that when
redirecting to index.

It also updates the views to always include those parameters.