Associate elements with page version
Element do not belong to a page anymore, instead they belong to a version of a page.

Therefore lots of page elements relations needed to be rewritten to be either through page's public_version or current_version.

Page#elements has been aliased as public_elements to not break existing sites.