Render the admin page tree on the client side
This change introduces Handlebars as a client side templating language
in order to speed up the rendering of the page trees in the admin
interface. This will result in significant speedups in both the pages
overview and the link dialog.

It also introduces a new type of serializer and API that enables quick
retrieval of the page tree from any node.