Ensure Element contents to be at least ordered by id
Element's contents are ordered by position, but the position of contens is scoped by its essence type. So different contents can have the same position on the same element.

In order to be sure we always talk about the same content, if we ask for the first one (ie. in `Element#preview_text`), we additionally order the contents by id.