Redirect to page sites host when visiting from Admin
Currently, with a multi-site setup, when visiting a page from
the edit page view in the admin, the admin user gets redirected to
a page's path - however, every page has a site, too, and that site
should be visited so that we don't get 404s.

If the page's site is a catch-all ("*"), we just redirect to the pages
`path`.