Move `current_ability` into it's own module.
A common need is to render alchemy content within the another context
(the app or another engine). An example of this would be the alchemy-solidus
gem. We do this by included `Alchemy::ControllerActions` into the controllers
for the context we want to use the Alchemy content in.

The problem is when this is done IF that gem is also using cancancan (like
Solidus does) then the `current_ability` method from Alchemy can end up
overwritting the `current_ability` of the controller doing the including.

By moving `current_ability` to it's own module it allows the including
controller to decide if that method is desired or not.