Fixes use of Alchemy::Resource with namespaced models
When using a namespaced base model for an Alchemy module,
there used to be confusion as to how to name paths. This commit
assumes that the models namespaces will be used also in the controller
world, ie: If your model is "MySite::Party", your resource controller
will be called "Admin::MySite::PartiesController", and the respective path
helpers will be called "admin_my_site_parties_path" (and so on).

The Generator is not yet up to date with this.

Fixes #729