Introduce automatical module scope filters in admin view
When a modules resource provides a `alchemy_resource_filters` class
method providing an array of strings, this will automatically show
a select filter with the strings as options that link to according
scopes in the module admin view.

**Example**
Given a module resource Event with the following code:

    class Event < ActiveRecord::Base
      def self.alchemy_resource_filters
        %w(released unreleased)
      end

      scope :released, -> { where(released: true) }
      scope :unreleased, -> { where(released: false) }
    end

The module admin view will show a select filter with the values
_all_, _released_ and _unreleased_ that link to the index page
with the according scope applied.