Redirect to instead of render public child of root
If requesting the root url and this is not public and `redirect_to_public_child` is set to true (default), we should redirect instead of rendering the page.

Also fixes a bug where requesting a non public page with non public children and disabled `redirect_to_public_child` never raised a 404.

Fixes #1014