Ditch bourbon button and use own mixin
It was very hard to override and customize the Bourbon buttons to Alchemy buttons styles. So it's more easy to use our own buttons that we have already and don't rely on external mixins.

Also the Bourbon buttons are deprecated and will be removed in the next mayor version anyway.