Store rendered pictures in public folder
In order to restore the current behavior we store the rendered picture in the public folder, so web servers can pick this up without even hitting the Rails app. This only happens when caching is enabled, though.

As we plan to copy the dragonfly config into the host app while installing, this can be configured to the need of the developer, as one may want to cache the images differently (Redis, S3, memcached, etc. pp)