Only update expanded elements with essence selects
If we save an element we need to update elements that have essence selects in them, because maybe this select references the updated element. A very rare case, but unfortunately we need to take care.

In order to make this fast and only when these conditions are met:

1. The element needs to be on the same page
2. The element needs to be expanded (not folded)
3. The element must not be trashed
4. The element needs to have at least one essence select

I hope we can remove this in a future version, but without a reference to an element we can't know if an element needs to be updated or not.