Fix factory fixtures for use in host app
When using the alchemy factories in tests in the host app, the
attachment and picture factory couldn't be loaded, as they referenced the
image.png that was located in the spec/fixtures folder. The spec folder
isn't included in the packed gem, though. This adds the image to the
test_support path.