Remove custom image caching
The image caching solution proposed in the generator leads to
errors with complex jobs, as the job serialization generates directory
names > 255 characters, which exceeds the limits of many popular OSs.

Instead, we leave caching up to the developer.

This commit also removes the Picture ID from the URL, as that was only
included to organize files.