Collapse main menu on small screens
This uses Javascript to collapse the main menu if the screen is smaller than 1024px. This is not ideal, but is way better than rewriting tons of sass selectors in the way, that they support @media queries and normal css selectors. It's just not possible without a mayor rewrite.

The downside is, that if one makes a full page reload, the main menu first appears in the full width and then shortyl after collapses. I think, that is something we can live. We have Turbolinks page transitions most of the time anyway. I doubt any user will ever notice.

The advantage is, that we could now add a navigation toggle button with not much additional effort.