Fixes the locked page notice in sitemap
Since the template gets compiled once into a Handlebars template, the only page that gets hit by ruby on template compile time is the root page. The sitemap renderer replaces all accurances of this id with the actual page id [1] to make the links work.

The "page locked by" notice also only gets rendered once, but since this is always the root page it happens that all other pages have the name of the user on all other pages as well (or 'unknown', if the root page is not locked).

This fixes it by putting the "page locked by" notice into the serializer.

[1] https://github.com/AlchemyCMS/alchemy_cms/blob/5dc29f5b88ee9f7dece474dc2c1de25b12204df7/app/assets/javascripts/alchemy/alchemy.sitemap.js.coffee#L15-L16