Remove tinymce editor instances before initializing
We have several places where we need to reinit tinymce editor instances. Sometimes they are already initlialized. Instead of manually removing them in all of these places, we remove the instances before initializing it.

This should fix lot of UI issues, where the tinymce does not get initialiazed correctly. For instance if an element has expanded nested elements, but is itself collapsed.