Refactor uploader specs for attachment and picture
Previously we made the silly assumption that nil values
passed as files lead to invalid uploads.

This works in Rails 4 as the params from test controllers
won't be converted into strings, as they would in real controllers.

In Rails 5 this is fixed, so that nil values passed as params
will correctly be converted into strings even in controller tests.

Now we make a correct invalid file by uploading in unsupported file type.