Future proof option params parsing
Previously we dealed with Strings, Hashes and Arrays (which are obviously never used, because we symbolised keys in all cases).

Now only distinguish between JSON strings and ActionController::Parameters, everything else is never used and have a Hash fallback for emtpy strings and nil.

By not asking explicitly for Hash anymore we are future proof as Rails 5 doesn't inherit from Hash for ActionContoller::Params anymmore.