Don't convert AR relations to arrays in breadcrump
Previously we converted the active record relation for pages
in the breadcrump helper into arrays to support options like
`:reverse` and `:without`.

Now we use the active record counterparts `#reorder` and `where.not` instead.
We could cache the breadcrumb views now if someone wanted to.

Also the implementation used in-self-replaced calls (`#reverse!` and `#delete`)
that where called on new arrays through `#to_a`, which lead to bugs in Rails 5.