Only allow floats and ints as fixed ratio for crop
Former we only checked if a fixed ratio was passed or not by checking its existence.
In Rails 5 all param values are strings, even booleans. This leads to NaN errors.
By checking the value to match an integer or float string we make sure this does
not happen any more.