Raise while creating default language in site model
Since we now create the default site the time we try to load it the first time and the site creates it’s own default language, we don’t need to check for the default language anymore.

Instead we raise an error if the configuration for default language can’t be found in the site model while we create the default language for that site.