Fix specs for MySQL date precision
MySQL has a complicated relationship with date types. Specifically,
according to the docs their default precision for storing datetimes is 0
(i.e. 1-second resolution). This is contrary to the SQL standard of 6
digits of sub-second precision. Because of this, many tests that set and
update date columns within the same second present false negatives. We
work around this by using the Timecop library to enforce super-second
resolution on the tests where needed.