Make essence delegation tests independent of hard-coded fixtures
Before this change the delegation tests in the shared essence examples relied on an element with a certain name and a content with a text essence.

When testing an essence in an external engine for Alchemy I don't want to be forced to look into the Alchemy specs to understand I have to set up very certain fixtures that have nothing to do with my add-on I am building.

This change allows external engines to better test essences with the shared examples provided by Alchemy.