Convert element definitions into hash with indifferent access
The element definitions read from the elements.yml file holds hashes with strings as keys. As we sometime pass action controller params (that have indifferent access as well) and sometimes just hashes (most of them in tests) we can't be sure to always have strings as keys.

As it actually doesn't matter if the hash keys are strings or symbols we should use a HashWithIndifferentAccess.

That will help us get rid of stringify_keys and symbolize_keys we use all over the place and is deprecated in Rails 5 and will be removed from Rails 5.1