Do not symbolize nor stringify hashes in content factory
As the element definitions hash has now indifferent access we can use symbols all the time and don't need to convert the hashes anymore.

We could stay with strings, but this would lead to lots of changes in the test suite where we use lots of symbols instead of strings for hash keys.

And as ActionController::Parameters (the true source of theses hashes) have also indifferent access this actually doesn't matter.