Do not symbolize attributes hash while creating element
Instead we create a new hash from sanitized parameters.

Rails keeps the indifferent access from the controller params when converting into a hash, so we can be sure that accessing keys via symbols and strings work.

This also removes a single line private method that was only used from another private method that gets rid of the element name in the parameters. We do this inside the calling method instead.