Fix n+1 when rendering associated resources
Prior to this commit, Alchemy would execute an ActiveRecord
find for every resource item in a list. However, due to the resource
controller preloading associations, many times this association will
already be in memory, so that sending the association name should be
much faster.

This also refactors the method somewhat, and it adds a spec for rendering
"Not found" if the association is empty.