Resource Controller: Allow overriding authorize_resource
In a host app, I want to use scopes as resource models for index actions.
Scopes can not be authorized through CanCanCan like actual models can.

This is a pretty remote use case, so I won't try baking that into the
already pretty complex resource handling system. However, it would be
convenient to be able to override the authorization logic from my host app.
This commit achieves that.