Do not ask for systempage everytime we load the page definition
While rendering the sitemap we ask for the page definition.

Previously we immediately returned an empty hash if the page is a `systempage?`.
That causes two extra queries for each page we render.

Instead we do not ask for `systempage?` anymore, because if the page definition
can't be found we already return an emtpy hash.

This speeds up rendering large sitemaps by about 6 times.