Handle custom errors in Alchemy::Picture#url
Currently we raise custom errors in the Alchemy::Picture#url
method without rescueing them later. This causes pages and the image library
to fail completely if one the images could not be rendered because the image file
is missing or the image format could not be used.

This is bad user experience. We should not display the broken image and not break the whole page.