Remove display cancan alias
The `display` cancan alias is also defined by Spree/Solidus. And because Spree/Solidus clear all pre defined aliases in their Ability class it can happen that our alias is not working anymore. Their alias is then used instead.

As we actually do not need this alias anywhere except in the permissions definition we can just remove it.