Make EssencePicture url options a HashWithIndifferentAccess
Content's settings - which we use to merge with the default EssencePicture `picture_url` options - are a `HashWithIndifferentAccess`. But the default url options were a simple `Hash` instead. Merging these two leads to url options having the `format` and `size` keys duplicated once as `String` once as `Symbol`. By using a `HashWithIndifferentAccess` for the `picture_url_options` we fix that.