Make EssencePictureView DEFAULT_OPTIONS a HashWithIndifferentAccess
Content's settings - which we use to merge with the EssencePictureView default options - are a `HashWithIndifferentAccess`. But the default options were a simple `Hash` instead. Merging these two leads to url options having the `format` and `size` keys duplicated once as `String` once as `Symbol`. By using a `HashWithIndifferentAccess` for the `DEFAULT_OPTIONS` we fix that.