Do not crop picture if crop option is set to false or nil
Only if the crop option is set explicitly to true we crop the image, even if `crop_from` or `crop_size` values are given.

This is a change to current behavior where the image was always cropped if `crop_from` or `crop_size` values were given, even if `crop` was set to `false` or left `nil`. This is a very confusing behavior.

Fixes #934