Convert preview window js code into vanilla js
We used to rely on jQuery being present in the preview window. If it's not present we loaded jQuery via our custon jQuery loader. This removes support for IE (even 11), but we never officially supported IE anyways and Edge should work as any other modern browser.

We can now remove the custom jQuery loader and lower the page load we inject into the preview window dom.