Only scroll to element if focused from preview
With the former change of focusing the element on every click we get issues if the element body is longer than the viewport. We had this problem before, but as we now focus the element on each click this is getting more annoying.

Solved by only scrolling the element editor into view when focusing the element from within the preview frame.