Add a compact nested element style
This compact element is great for single content elements (like pictures in a gallery). It uses the half width of the normal element and has an overall densed style to safe space in elements with lots of nested elements.

Define a compact element by adding `compact: true` to its element definition.