Do not drop cells table after they got migrated
Dropping the cells table in a migration is wrong as the migrations run before the Alchemy upgrader runs.

Leaving the table and column is not harming anyone and deleting users data is not a good idea anyway. Whoever dislikes the table sitting around can generate a migration to remove it.