Expose the element into partials as local object
To be consistent with Rails' collection partial renderer we also expose the element as local object named after the element into the view partials.

## Example

Given an element named "article" and a `article_view` element partial, you can now access the element as `article_view` local variable as well as the `element` variable.