Use Enumerable to find Element Contents
Prior to this commit, `contents_by_type` and `contents_by_name` would
always do an SQL query. With this commit, they can re-use the
`@element.contents` array if it's already loaded.

There is a good chance this does not do that much, and it does not
reduce the amount of SQL queries in the Element spec. I don't feel it
makes sense to test this explicitly either, as the functionality does
not change one bit - only the speed of execution should (in some cases)
improve.