Deprecate options_from_params admin controller method
Passing around options in requests from essence views into admin
controllers is deprecated and will be removed in Alchemy 5.0

This was only necessary because essence editor might have local
options that needed to be preserved over requests.

Local options in essence editor partials are not supported anymore,
therefore we do not need to pass options around as well.