Implement attach page to menu feature
A page can now be attached to a menu.

This works with the "visible in navigation" checkbox, but only
if Menus are already present - it will fall back to old behavior
instead.

Once attached to a menu node it you cannot detach it from the page,
you need to remove the menu node first.