Add ingredient associations
An ingredient association is a way to be able to eager load
associated records of essence classes. The EssencePicture for example
`belongs_to :picture`.

In able to eager load we create an `belongs_to :ingredient_association`
for you and add an ActiveRecord hack that skips eager loading on records
not having such relation.

That way we can load a page and include all elements, their contents,
essences and their associated records as well.

This should speed up admin pages with large element lists and API calls
to pages with many elements a lot.