Fix etag spec assumption
Rack 2.2.0 sends an etag even if resource should not be cached.

This is ok since an etag is just a digest of the body and Rails
also sends a no-cache Cache-Control header for resources that must
not be cached.

Removing the test since we must not test internal Rack/Rails behavior.