Add default sort order to resources
Before we where relying on the default sort order of the database.
That is, but does not have to be, the primary column in most databases.

This changes this behavior to introduce a default_sort_order that
uses the name column (if the resource has one) or the first attribute
in an ascending sort order.

Also adds a nice little indicator to the sorted column.