Remove implicit Site and Language creation (#1763)
We do not want to implicitly create a site when none or a language when none is there. This
should be in control of the admin user.

Instead, an admin user starting a new project gets redirected to sites/languages whenever she tries accessing the pages, layout pages, or menu admin.

This will make it easier to explicitly do test setup for multi-site and multi-language projects.