On Boarding Flow (#1770)
* Rename LanguageNeededRedirect module

We will add more functionality to that module

* Only redirect to languages form if current language is missing

We do not need to test for any language to be present if we load the current language and only redirect to the new language form if that is missing.

* Only redirect to new site form if current site is missing

* Show new site form if no site present yet

Since we do not automatically create new sites implicitely this will help onboarding new users.

* Show new language form if no one is present

Now that we do not implicitely create the default language anymore this will help onboarding new users.

* Redirect to languages admin after creating site

A site needs languages. To better onboard new users we should redirect to the language admin after creating a new site (that does not have any languages yet REMOVE ME)

* Remove border from panel

* Add link style for hints

We want to see links in hints

* Fix languages redirect controller spec

* Prevent site from being destroyed if languages attached

We would either lose lots of data or orphaned records.

* Prevent language from being destroyed if pages are attached

We must not delete a language if there are still pages present. We either would lose a lot of data or orphaned records.

* Refine new menu page

* Redirect to pages admin after creating language